. tools/Makefile.am, krbparse.py, rpcfg.py, updk5cfg.py, updwscfg.py: moved kerberos related function to a new file updk5cfg.py, and fixed some bugs in the execution path of UpdateAuthSettings.
